feat: Migrate money fields to long type in database schema
- Added migration to change money fields (GrandTotal, TaxAmount, Subtotal, ShippingFee, DiscountAmount) from decimal to long in fulfillment_saga_states and checkout_saga_states tables.
- Updated OrderDbContextModelSnapshot to reflect changes in the database schema.
- Modified CheckoutSagaState and FulfillmentSagaState to use long for monetary values.
- Introduced new FulfillmentSagaState and its entity configuration for handling fulfillment processes.
- Enhanced GetCheckoutStatusQueryHandler to throw NotFoundException if the saga is not found.
- Updated OrderInfrastructureExtension to register event publisher services.
- Adjusted Login page to handle registration intent from authorization context.
- Changed DefaultRedirectUrl in appsettings.json for UserService.

+++ libs/HiveSpace.Infrastructure.Messaging.Shared/CheckoutSaga/Events/PackageRejected.cs
@@ -6,5 +6,5 @@ public record PackageRejected
     public Guid    OrderId       { get; init; }
     public Guid    PackageId     { get; init; }
     public string  Reason        { get; init; } = null!;
-    public decimal PackageAmount { get; init; }
+    public long    PackageAmount { get; init; }
 }
